move loadAccounts to substrate example
