+++ examples/substrate-react-example/src/utils/index.ts
@@ -1,0 +1,2 @@
+export { loadAccounts } from "./loadAccounts";
+export { retrieveChainInfo } from "./retrieveChainInfo";
