+++ packages/sdk/src/chains/Substrate/types/index.ts
@@ -6,14 +6,5 @@ export type XcmMultiAssetIdType = {
     };
   };
 };
-export type SubstrateConfigAssetType = {
-  assetName: string;
-  assetId: number;
-  xsmMultiAssetId: XcmMultiAssetIdType;
-};
-export type SubstrateConfigType = {
-  domainId: string;
-  appName: string;
-  provider_socket: string;
-  assets: SubstrateConfigAssetType[];
-};
+
+
