+++ packages/sdk/src/chains/Substrate/utils/index.ts
@@ -3,8 +3,6 @@ export { getAssetBalance } from './getAssetBalance';
 export { getBasicFee } from './getBasicFee';
 export { getNativeTokenBalance } from './getNativeTokenBalance';
 
-export { loadAccounts } from './loadAccounts';
-export { retrieveChainInfo } from './retrieveChainInfo';
 export {
   deposit,
   throwErrorIfAny,
